recommit: revisit the become leader and become follower state change operations using V3 design; patched by Yang Ye; reviewed by Neha Narkhede and Jun Rao; kafka-343
git-svn-id: https://svn.apache.org/repos/asf/incubator/kafka/branches/0.8@1368092 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/scala/kafka/admin/AdminUtils.scala
@@ -91,10 +91,10 @@ object AdminUtils extends Logging {
     topics.map { topic =>
       if (ZkUtils.pathExists(zkClient, ZkUtils.getTopicPath(topic))) {
         val topicPartitionAssignment = ZkUtils.getPartitionAssignmentForTopics(zkClient, List(topic).iterator).get(topic).get
-        val sortedPartitions = topicPartitionAssignment.toList.sortWith( (m1,m2) => m1._1.toInt < m2._1.toInt )
+        val sortedPartitions = topicPartitionAssignment.toList.sortWith((m1, m2) => m1._1 < m2._1)
 
         val partitionMetadata = sortedPartitions.map { partitionMap =>
-          val partition = partitionMap._1.toInt
+          val partition = partitionMap._1
           val replicas = partitionMap._2
           val inSyncReplicas = ZkUtils.getInSyncReplicasForPartition(zkClient, topic, partition)
           val leader = ZkUtils.getLeaderForPartition(zkClient, topic, partition)
